Check if a binary tree is a binary search tree
A binary tree is a tree that has nodes that have at
the most, two children.

A binary search tree is a tree such that its nodes
have at the most, two children, and the left child's
value is less than or equal to the root's value, and
the right child's node is greater to the root's value.

This is true for all nodes as we travel down the tree
on the left and the right.